Afficher détails produit

+++ src/main/kotlin/org/ldv/groovium/model/dao/ProduitDAO.kt
@@ -2,5 +2,5 @@ package org.ldv.groovium.model.dao
 import org.ldv.groovium.model.entity.Produit
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ProduitDAO : JpaRepository<Produit, Long> {
+interface ProduitDAO : JpaRepository<Produit, Int> {
 }
